+++ clients/vscode-hlasmplugin/src/constants.ts
@@ -14,5 +14,6 @@
 
 export const pgm_conf_file = 'pgm_conf.json';
 export const proc_grps_file = 'proc_grps.json';
+export const bridge_json_file = '.bridge.json';
 export const hlasmplugin_folder = '.hlasmplugin';
 export const ebg_folder = '.ebg';