+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Source code colorization may be flickering while typing
 - Configuration request sent before initialization is done
 - VSCode enters an infinite loop of opening and closing files
+- "pgm_conf.json not found" prompt shows up even when `.bridge.json` exists
 
 ## [1.7.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.6.0...1.7.0) (2023-03-08)